requirements.txt
PySide6~=6.5.0
python-docx~=0.8.11
